+++ packages/app/src/backend/requests/upload-device-file/upload-device-file.request.ts
@@ -8,7 +8,7 @@ import createEndpoint from "Backend/endpoints/create-endpoint"
 import { IpcRequest } from "Common/requests/ipc-request.enum"
 import DeviceResponse from "Backend/adapters/device-response.interface"
 import { UploadFilePayload } from "Backend/device-file-system-service/device-file-system-service"
-import { arrayBufferToBuffer } from "App/files-system/helpers"
+import { arrayBufferToBuffer } from "App/file-system/helpers"
 
 export interface UploadFileUIPayload extends Omit<UploadFilePayload, "data"> {
   data: Uint8Array
