+++ packages/app/src/file-system/services/crypto-file-service/data/image.jpg
