+++ packages/app/src/main/main.ts
@@ -25,13 +25,13 @@ import registerCopyFileListener from "App/main/functions/register-copy-file-list
 import registerWriteGzipListener from "App/main/functions/register-write-gzip-listener"
 import registerRmdirListener from "App/main/functions/register-rmdir-listener"
 import registerArchiveFilesListener from "App/main/functions/register-archive-files-listener"
-import registerReadFileListener from "App/files-system/listeners/read-file-listener"
+import registerReadFileListener from "App/file-system/listeners/read-file-listener"
 import registerGetApplicationConfigurationListener from "App/main/functions/register-get-application-configuration-listener"
 import registerGetFileDataListener from "App/main/functions/register-get-file-data-listener"
 import registerPureOsDownloadProxy from "App/main/functions/register-pure-os-download-proxy"
 import createDownloadListenerRegistrar from "App/main/functions/create-download-listener-registrar"
-import registerEncryptFileListener from "App/files-system/listeners/encrypt-file-listener"
-import registerDecryptFileListener from "App/files-system/listeners/decrypt-file-listener"
+import registerEncryptFileListener from "App/file-system/listeners/encrypt-file-listener"
+import registerDecryptFileListener from "App/file-system/listeners/decrypt-file-listener"
 import registerOsUpdateAlreadyDownloadedCheck from "App/main/functions/register-os-update-already-downloaded-checker"
 import {
   registerDownloadHelpHandler,
