+++ packages/app/src/restore-device/actions/start-restore-device.action.test.ts
@@ -21,12 +21,12 @@ import { DeviceState } from "App/device"
 import { StartBackupDeviceError } from "App/backup-device/errors"
 import uploadDeviceFile from "Renderer/requests/upload-device-file.request"
 import startRestoreDeviceRequest from "Renderer/requests/start-restore-device.request"
-import readFile from "App/files-system/requests/read-file.request"
-import decryptFile from "App/files-system/requests/decrypt-file.request"
+import readFile from "App/file-system/requests/read-file.request"
+import decryptFile from "App/file-system/requests/decrypt-file.request"
 import { waitUntilRestoreDeviceFinished } from "App/restore-device/helpers"
 
-jest.mock("App/files-system/requests/decrypt-file.request")
-jest.mock("App/files-system/requests/read-file.request")
+jest.mock("App/file-system/requests/decrypt-file.request")
+jest.mock("App/file-system/requests/read-file.request")
 jest.mock("Renderer/requests/upload-device-file.request")
 jest.mock("Renderer/requests/start-restore-device.request")
 jest.mock("App/restore-device/helpers/wait-until-restore-device-finished")
